Don't allow skip value of Neo query to be negative number
In practice this won't happen as the endpoint has code to set
sensible default value. However using the ConceptService directly
allows you to set such params that the skip number becomes negative.
This is actually true for the integration tests.
Apparently that negative values are problem for Neo 4.3.6 but
were not a problem for 3.5.